+++ src/index.ts
@@ -2,7 +2,7 @@ export { define } from './define/index.js';
 export { parent } from './service/index.js';
 export { pool } from './pool/index.js';
 export { ConcurrencyOption } from './types/pool.js';
-export { Messenger, messages } from './messenger/index.js';
+export { Messenger, messages, messengers } from './messenger/index.js';
 export { ServiceCluster } from './service_cluster/index.js';
 
 export type { TaskDefinitions } from './types/definitions.js';