+++ src/messenger/index.ts
@@ -1,3 +1,3 @@
 export { Messenger } from './messenger.js';
 export { isMessengerTransferObject } from './utilities.js';
-export { messages } from './messages.js';
+export { messages, messengers } from './messengers.js';