+++ src/messenger/messengers.ts
@@ -64,6 +64,7 @@ function seek() {
 }
 
 /**
+ * @deprecated Use `messengers` instead.
  *
  * An object containing functions to be used within workers when interacting with {@link Messenger}s.
  *
@@ -73,3 +74,14 @@ function seek() {
  * messenger.onMessage<string>((data) => console.log(data, 'received!'));
  */
 export const messages = Object.freeze({ use, seek });
+
+/**
+ *
+ * An object containing functions to be used within workers when interacting with {@link Messenger}s.
+ *
+ * @example
+ * const messenger = await messengers.use('foo');
+ *
+ * messenger.onMessage<string>((data) => console.log(data, 'received!'));
+ */
+export const messengers = Object.freeze({ use, seek });