+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Docs for `closeAll()` and `setRef()` under "Using Messenger".
 - `launch()` method to `ServiceCluster`.
 - Docs for `launch()`.
+- `messengers` object (identical to `messages`).
 
 ### Changed
 
@@ -24,6 +25,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Deprecated
 
 - `launchService()` method on `ServiceCluster`.
+- `messages` object in favor of new identical `messengers` object.
 
 ## [0.1.3] - 2022-02-11